+++ include/sphericalSampler.h
@@ -8,21 +8,20 @@ class SphericalSampler {
  private:
   std::default_random_engine re;
 
-  Vec3 getVelocity(Vec3 pos, Vec3 center, double rb, double mb, double rd, double md, double G);
+  Vec3 getVelocity(Vec3 pos, Vec3 center, float rb, float mb, float rd, float md, float G);
 
-  double sampleFromLinear(double rd);
+  float sampleFromLinear(float rd);
 
-  inline double implicitInvCDFLinear(double r, double rd, double cdf) {
-    return 3 * std::pow(r, 4) - 4 * rd * std::pow(r, 3) + cdf * std::pow(rd, 4);
+  inline float implicitInvCDFLinear(float r, float rd, float cdf) {
+    return 3 * std::powf(r, 4) - 4 * rd * std::powf(r, 3) + cdf * std::powf(rd, 4);
   }
 
-  inline double implicitInvCDFLinearDer(double r, double rd) {
-    return 12 * std::pow(r, 3) - 12 * rd * std::pow(r, 2);
+  inline float implicitInvCDFLinearDer(float r, float rd) {
+    return 12 * std::powf(r, 3) - 12 * rd * std::powf(r, 2);
   }
 
  public:
   SphericalSampler();
 
-  std::vector<Vec3>
-  sample(Vec3 center, double rb, double mb, double rd, double md, double G, int n);
+  std::vector<Vec3> sample(Vec3 center, float rb, float mb, float rd, float md, float G, int n);
 };
