+++ source/main.cpp
@@ -15,17 +15,17 @@ int main() {
   */
   const int n = int(5e4);
   Vec3 galaxyCenter = Vec3(30, 30, 30);
-  double rb = 3;
-  double mb = 15;
-  double rd = 15;
-  double md = 60;
-  double thickness = 0.3;
-  double G = 4.5e-3;
+  float rb = 3.0f;
+  float mb = 15.0f;
+  float rd = 15.0f;
+  float md = 60.0f;
+  float thickness = 0.3f;
+  float G = 4.5e-3f;
 
   DiskSamplerLinear diskSampler;
   std::vector<Vec3> state = diskSampler.sample(galaxyCenter, rb, mb, rd, md, thickness, G, n);
 
-  std::vector<double> masses(n, md / n);
+  std::vector<float> masses(n, md / n);
   auto externalField = [galaxyCenter, rb, mb, G](Vec3 pos) -> Vec3 {
     return externalFieldBulge(pos, galaxyCenter, rb, mb, G);
   };
@@ -34,9 +34,9 @@ int main() {
   int dims[] = {gridPoints, gridPoints, gridPoints};
   KissFFTAdapter<float> fftAdapter(dims, 3);
   Grid grid(gridPoints, fftAdapter);
-  double effectiveBoxSize = 60;
-  double H = effectiveBoxSize / (gridPoints / 2);
-  double DT = 1;
+  float effectiveBoxSize = 60;
+  float H = effectiveBoxSize / (gridPoints / 2);
+  float DT = 1;
 
   PMMethod pm(state, masses, externalField, H, DT, G, InterpolationScheme::CIC,
               FiniteDiffScheme::TWO_POINT, grid);
